Bryan Ramos y Milagors Escorcia Cambios 11/04/2023

+++ EL/Usuario.cs
@@ -28,7 +28,7 @@ namespace EL
         [Required]
         public bool Bloqueado { get; set; }
         [Required]
-        public byte IntentosFallidos { get; set; }
+        public short IntentosFallidos { get; set; }
         [Required]
         public int IdRol { get; set; }
